refactor: move db code to its own file
